git: patch-id
